Fix schema missing for half of the global preferences
The global preferences were missing tooltip and validation for all
EditorSettings. This was due to schema for "auto_complete_triggers"
not matching (due to missing the "rhs_empty" property in schema) and
thus the whole "anyOf" schema becoming invalid. Since that item of anyOf
became invalid, the other was tested and matched (because
"additionalProperties" are allowed).

Use "allOf" instead which works correctly for this case.