Include schemas that only have uri
Those schemas were skiped, beacuse thay don't have fileMatch. But we need those.

If we want to use the 2019-09 (ex draft-8) of the spec.
Previously we didn't included it because it didn't had a fileMatch.

now we can assign the 2019-09 draft to a schema with "$schema" keyword.

"$schema": "https://json-schema.org/draft/2019-09/schema"